- WIP: Using vitest (working) - WIP: Using ESM (fully) (run server ok) - Lowered verboseness of plugin loader logs.
vitest is working with both basic tests and e2e tests in heli-crash-broadcast.test.ts that use dynamic imports with await import that are loading plugins.
I think Jest could still work but as with vite, when adding `preset: 'ts-jest/presets/default-esm'` it stops on lodash not being ESM, vitest does that too. So lodash-es it is.
`"type": "module",` is not necessary for `run` but does break vitest on e2e test. It also makes sense to place it since project is (officially) ESM now. Before, I was kind of using `tsx` magic to make it work with `await import` while retaining some commonJS syntax like `__dirname`.
`__dirname` is now replaced by ESM equivalent.

- jest imports have been replaced by drop-in vitest.
- vitest is said to be faster than jest, syntax is almost the same. If there is any issue we can probably switch back to Jest.

Note: file with import.meta.url will work fine, but Webstorm will complain, this is because file will have to be renamed to .mts (in another commit)

Note: e2e is still failing, but because import path is wrong (`src/config`), this is somewhat expected, and will be fixed a way or another.